+++ CHANGES.md
@@ -7,6 +7,7 @@ History
 * Change `Frame.executionContest` from property to coroutine
 * Implement `Frame.content` and `Frame.setContent` methods
 * `page.tracing.start` accepts custom tracing categories option
+* Add `Browser.process` property
 
 
 ## Version 0.0.14 (2018-03-14)